requirements.txt
coverage==4.0.2
pytest==2.8.2
pytest-cov==2.2.0
pytest-flake8==0.2
pytest-mock==0.9.0
requests==2.10.0
